parquet: Update benchmark to reuse generated block (#12594)
This PR is a follow up to https://github.com/grafana/mimir/pull/12532 

The previous PR was able to benchmark store-gateway current performance
against the newly created parquet code however it would always have to
generate the data before each run, slowing us down if we wanted to test
with different configurations or block sizes.

In this PR:
- There is a new `blockgen` tool we can use to generate a tsdb + block
directory for an user. Includes README documenting the opts.
- The existing benchmark now accepts a `-benchmark-tsdb-dir` argument
accepting an entire tsdb. If set it won't generate data on the fly for
the benchmark, if not set then it will.

Quick run:
```
$ cd pkg/storegateway/parquetbench
$ go run blockgen/main.go -zones 1000 -instances 10 -verbose -output benchmark-data-15M # generates a block with 15M series and 1k cardinality on zones
...
$ go test': go test -bench=. -benchmark-store=parquet  -benchmark-tsdb-dir=./blockgen/benchmark-data-15M 
...
```

Note: Remember to use `-benchmark-store=parquet` to benchtest the
parquet code and `-benchmark-store=tsdb` for the current store-gateway
code.