drop 'grafana' in 'grafana mimir' product name on landing page (#12601)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

This pull request makes a minor update to the documentation for Grafana
Mimir by changing the hero title from "Grafana Mimir" to simply "Mimir"
in the `docs/sources/mimir/_index.md` file. This is part of an effort to
obtain TM protection for "Mimir".

#### Which issue(s) this PR fixes or relates to

Fixes https://github.com/grafana/website/issues/27234

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.