index planning: Disable planning for blocks older than 3 hours (#12608)
## Summary
For blocks older than 3 hours, disables lookup planning to avoid using
outdated per-block statistics

## Why
When the ingester stops receiving writes it's in memory state, on which
we base statistics is empty. This leads to very suboptimal plans.

The 3-hour threshold accounts for:
- 1h: data in memory (head block) 
- 2h: typical compaction window which is still kept in memory


This is obviously suboptimal and has problems for when the data pushed
is far away from `now()`, but it's a good enough approximation to
unblock using index lookup planning.

## Next steps

A more permanent solution is per-block stats
pull/12607