[cost attribution]: reduce max cost attribution label cardinality to 2000 (#12625)
#### What this PR does

This PR reduces the default value of the
`-validation.max-cost-attribution-cardinality` limit from `10K` to `2K`.
This change is needed to reduce pressure on distributors and ingesters,
and to preventing performance degradation.

#### Which issue(s) this PR fixes or relates to


#### Checklist

- [ ] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Yuri Nikolic <durica.nikolic@grafana.com>
Co-authored-by: Andy Asp <90626759+andyasp@users.noreply.github.com>