docs: update get-started-helm-charts replacing the deprated meta-monitoring (#12623)
#### What this PR does

Here I'm updating the `get-started-helm-charts` docs, replacing the
steps that use the deprecated meta-monitoring with a step that installs
Grafana Alloy. Even though this is more verbose than it used to be, this
represents the users' experience better.

Also, updated the doc's bits to list the actual list of Mimir pods,
which now include Kafka by default.

#### Which issue(s) this PR fixes or relates to

Relates to https://github.com/grafana/mimir/issues/5860

#### Checklist

- [ ] Tests updated.
- [x] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>
Co-authored-by: Nick Pillitteri <56quarters@users.noreply.github.com>