continuous-test: Add new test for verifying ingest-storage record correctness (#12500)
#### What this PR does

Adds a new, off-by-default, continuous test that connects to the
ingest-storage Kafka and verifies ingest-storage records. It intercepts
a random subset of live write requests to evaluate each time the
continuous tester runs.

Its primary purpose is to test the new V2 record format. If the cell is
using the V1 record format, this test marshals records to V2 and asserts
that both record formats map to the same actual request. It also emits
statistics about the difference between V1 and V2, for the purpose of
observing how V2 behaves without putting the write path at risk.

Once the V2 format reaches stability, it becomes less relevant - but it
can be a nice utility for cells wishing to migrate to V2, and also the
general idea of the test might be applicable to other areas and could be
extended or modified going forward.

The test does not alert when it fails, at least for now. It simply logs
its output and continues.

It is designed to only apply a bounded and configurable amount of
pressure to Kafka. It consumes up to a fixed amount of records per run,
regardless of the throughput of the cell, and manually commits offsets
to "seek" forward in time when limits are exhausted. Most records are
never fetched, and the number of fetches and commits performed is a
function of the limit.

#### Which issue(s) this PR fixes or relates to

contrib https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.