ruler: Refactor GroupLoader to use the same filesystem abstraction as the mapper (#12617)
#### What this PR does

The `mapper` interacts with the disk through a filesystem abstraction,
https://github.com/spf13/afero. The mapped files are read by other
components, notably the `GroupLoader`, which don't use the abstraction.

This PR expands use of the abstraction to fully cover the file
interaction in the ruler, and allows a common FS instance to be injected
among components.

We continue injecting an `OsFs` as before. Everything still interacts
with the real filesystem without a change in behavior, but through the
abstraction instead of directly.

This sets us up for a future replacement of the `afero.Fs` with an
in-memory data structure. It may be an in-memory implementation of
`afero.Fs` at first, or potentially a different thing altogether. The
eventual goal is to speed up rule group syncs, because `manager.Update`
acquires a lock that holds up ruler config API requests and causes
latency spikes. More context in
https://github.com/grafana/mimir-squad/issues/2336

#### Why Afero and not `io/fs`?

Afero is an older older tool that predates the `io/fs` package in the
standard library. However, `io/fs` notably only contains support for
reading the filesystem, not writing to it. Afero is a superset. The
`mapper`, which already uses afero, cannot migrate entirely to `io/fs`
due to the [writes and directory
creation](https://github.com/grafana/mimir/blob/7a01df44d6afacfe4bbdb8272db563242837b31e/pkg/ruler/mapper.go#L168)
that it does.

`afero.Fs` does not implement `io/fs.FS` either, but only due to Go not
supporting covariance in return types (for good reason) and not due to
missing functionality. It's possible to have part of the code like the
`GroupLoader` depend only on `fs.FS`, but mixing the abstractions
requires manually writing an adapter between `afero.Fs` and `fs.FS`
which doesn't seem worth it to maintain.

`afero` also ships with an in-memory FS, as a quick and easy way to
remove the disk interaction from the ruler. From docs it's intended for
[use-cases beyond
testing](https://github.com/spf13/afero?tab=readme-ov-file#example-2-caching-a-slow-filesystem),
and the [code](https://github.com/spf13/afero/blob/master/memmap.go) is
straightforward, so it seems like a reasonable choice.

#### Which issue(s) this PR fixes or relates to

Contrib https://github.com/grafana/mimir-squad/issues/2336

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.