query-frontend: enable labels-query-optimizer-enabled by default (#12606)
#### What this PR does

We've been running the `labels-query-optimizer-enabled` enabled in
Grafana Labs production for a couple of months, and are confident now to
make it the default. We still keep the feature-flag, just in case.

Note, I don't have much of context about the feature, and thus, I don't
mind keeping the flag "experimental" for the next release (cc
@pracucci).

#### Which issue(s) this PR fixes or relates to

Following #12080

#### Checklist

- [ ] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>