[r359] OTLP: Return HTTP OK when partially rejecting due to OOO exemplars (#12579) (#12667)
#### What this PR does

Modify the OTLP endpoint to return HTTP OK when partially rejecting due
to OOO exemplars, to adhere with
[spec](https://opentelemetry.io/docs/specs/otlp/#otlphttp-response).

I'm also modifying the ingester to include in gRPC error details whether an error is soft or not, so the distributor can recognize so-called soft errors (corresponding to partial rejection).

I've manually tested the fix in a development deployment, and can
confirm it works.

Relates to #12626.

(cherry picked from commit bf51551afb0474fd22fb4a9fd77dea445a863d1b)

#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Marco Pracucci <marco@pracucci.com>
Signed-off-by: Arve Knudsen <arve.knudsen@gmail.com>
Co-authored-by: Marco Pracucci <marco@pracucci.com>