Introduce operator specific Matcher value type
Create an operator specific Matcher value type to use instead of
Prometheus pointer labels.Matcher types so that we can use the
type as a key for maps and easily compare equality in tests. The
new type is converted to a Prometheus matcher before being passed
to a Queryable/Querier.