MQE: include details of query planning in traces (#12675)
#### What this PR does

This PR improves the traces emitted during the query planning process. 

Specifically, the original expression, optimised expression, original
plan and optimised plan will be added as events on the query planning
span.

I've opted not to add a changelog entry given this is not a user-facing
change.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Jon Kartago Lamida <jon.lamida@grafana.com>