MQE: perform query planning in query-frontends and send to queriers (#12551)
#### What this PR does

This PR introduces experimental support for performing query planning in
query-frontends, building on the querier changes in
https://github.com/grafana/mimir/pull/12302.

When enabled, query-frontends compute a query plan for each query and
send it to queriers for evaluation.

The key parts of this are:

* When enabled, query-frontends use the newly-introduced
`engineQueryRequestRoundTripperHandler` instead of the
`httpQueryRequestRoundTripperHandler` to handle instant and range
queries
* `engineQueryRequestRoundTripperHandler` is responsible for taking a
query request and passing it to the query-frontend's query engine for
evaluation
* The query-frontend's query planner is configured with the
newly-introduced `remoteexec.OptimizationPass` optimization pass that
injects a `RemoteExecution` node at the root of the query plan if the
query contains selectors but does not contain sharded or spun-off
expressions
* Keep in mind that the query-frontend's engine instance is used in two
contexts now: once while evaluating the overall sharded query in the
sharding middleware (if the query is shardable) and then while
evaluating the overall query (if the query isn't shardable) or each
individual leg (if it is shardable) in the roundtripper.
* This is why we don't insert `RemoteExecution` nodes in query plans
containing sharded expressions, as we don't want remote execution to run
during the evaluation initiated by the sharding middleware.
  * This is a temporary thing until sharding is running within MQE.
  * Similar logic applies to subquery spin-off as well.
* The `RemoteExecution` nodes are materialized as
`remoteexec.InstantVectorOperator`, `remoteexec.RangeVectorOperator` or
`remoteexec.ScalarOperator` operators.
* These operators use the frontend's `v2.RemoteExecutor` to submit
requests for evaluation which are then forwarded to query-schedulers and
then queriers for evaluation.

Subsequent PRs will introduce improvements or missing functionality,
such as:
* propagating strong consistency information to queriers
* sending query component flags (ingester-only, store-gateway-only or
both) to query-schedulers
* propagating query stats to query-frontends
* possibly batching up results from queriers

Out of scope for this PR, but will be introduced at some point in the
future:
* performing query sharding as part of the query plan
* performing subquery spin-off as part of the query plan
* performing range query splitting and caching as part of the query plan

In each of the cases above, the individual sharded legs / cacheable
extents are independently planned and evaluated.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [x]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.