Update mimir-prometheus to c2f03d6d14d2 for the OTLP endpoint rewrite (#12652)
This PR updates mimir-prometheus to
[`c2f03d6d14d2`.](https://github.com/grafana/mimir-prometheus/commit/c2f03d6d14d20b6f2f26cff2b8ce05c1ce3d7007)

Brings in https://github.com/prometheus/prometheus/pull/16951.

Changelog: OTLP: native support for OpenTelemetry metric start time to
Prometheus metric created timestamp conversion, instead of converting to
QuietZeroNaNs introduced in #10238. The configuration parameter
`-distributor.otel-start-time-quiet-zero` is therefore deprecated and
will be removed. Now supports start time for exponential histograms.
This is a major rewrite of the endpoint in upstream Prometheus and
Mimir. #12652

Implementation does away with the generated otlp translation code as we
don't need to rewrite it for `mimirpb` anymore. The storage logic is
behind an interface.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>
Co-authored-by: Arve Knudsen <arve.knudsen@gmail.com>