Calculate plan costs only once (#12658)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does
This PR precomputes costs for all plans once before sorting, so that
`slices.SortFunc` does not compute each plan cost every time it does a
comparison. This saves on CPU when using the `CostBasedPlanner`.

#### Which issue(s) this PR fixes or relates to

Loosely related to https://github.com/grafana/mimir/issues/11921

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.