Ensure query-frontend safely handles the case where a response is received multiple times for the same query (#12639)
#### What this PR does

This PR builds on https://github.com/grafana/mimir/pull/12551.

This PR fixes an issue where the `QueryResultStream` method on frontends
may panic or produce incorrect results if a response for the same query
is received multiple times.

I've also added a test to confirm that it correctly handles the scenario
where a response is received for a query ID that was used for a query
from a different tenant.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.