Refactor query sharding logic in preparation for MQE implementation (#12676)
#### What this PR does

This PR refactors some parts of the query sharding logic in preparation
for the MQE implementation of query sharding.

MQE's implementation of query sharding will re-use most of the existing
logic. It will run as an AST optimisation pass that wraps sharded legs
in a special function (likely named something like `concat` or
`sharded`), rather than embedding queries in a specially-constructed
vector selector.

To make this easier to implement, and reduce the size of the PR that
implements it, in this PR, I'm doing some refactoring of the existing
logic.

I've tried to break each change into its own commit for easier review.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.