blockbuilder: pre-initialize cortex_blockbuilder_scheduler_job_gap_detected (#12683)
#### What this PR does

Following https://github.com/grafana/mimir/pull/11867

In the mentioned PR I suggested adding the `partition` label to the
`cortex_blockbuilder_scheduler_job_gap_detected` metric. It has turned
out this creates more problems that benefits. I.e. because the counter
is initialized on the first increment, on scheduler's restart, this
creates a gap in the series. This isn't a problem typically. But for the
kind of counter that increments rarely — like the gap detector — this
creates a possibility for an alert to miss the increment (when the
counter starts with 1).

This PR updates the metrics

1. remove the partition label from the series
2. pre-initialize the possible "offset_type" series with 0

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>
(cherry picked from commit 00b427a42176ce1fc452697adf94a94d85c7c20c)