compactor: include sparse index headers in meta.json (#12416)
#### What this PR does
Tries to read sparse-index-header (best effort), includes in meta.json
if it exists. Skips the file silently if not found, fails for other
errors.

#### Which issue(s) this PR fixes or relates to
Fixes #12037 compactor: include sparse index header in meta.json

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: pranshu-raj-211 <pranshuraj65536@gmail.com>