query-frontend: Skip newly added integration tests that don't pass in GEM (#12699)
#### What this PR does

https://github.com/grafana/mimir/pull/12551 adds a new, experimental
query execution mode to query-frontends plus e2e tests for that mode.

The e2e tests don't pass in GEM, and interfere with GEM vendoring. Since
it's an experimental mode that nobody has enabled yet, we feel fine
keeping the code and temporarily disabling the tests.

#### Which issue(s) this PR fixes or relates to

n/a

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.