index planning: use planning for all sizeable blocks (#12696)
We had a patch (#12608) to prevent using stale stats. That wasn't
effective and still left the possibility of calculating stats on an
empty head, the head receiving series and then using the empty stats on
a non-empty head. This results in very inefficient plans. Now that we
have #12604, we can remove this logic.

This PR makes the logic for when we use planning a bit more testable
with the `lookupplan.PlannerFactory` and removed the changes from #12608

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>