MQE: report query stats for remotely executed queries (#12646)
#### What this PR does

This PR builds on https://github.com/grafana/mimir/pull/12551.

This PR adds reporting of query stats for remotely executed queries,
bringing the stats reported in line with those reported when remote
execution is not used.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.