storegateway: BucketChunkReader to validate chunk data length (#12693)
#### What this PR does

The PR adds a validation to the `BucketChunkReader` to make sure it
doesn't panic, if the parsed chunk length exceed the allowed buffer
length. See #12691 for the details.

#### Which issue(s) this PR fixes or relates to

Fixes #12691

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>