Block-builder-scheduler: Clean up initial probing and warn when we don't probe far enough. (#12701)
If the scheduler is started with a `max-scan-age` that is smaller than the actual age of the commit record, we will skip data. This change adds a warning log when this happens to make this situation more apparent.

And it polishes up the commentary around startup probing a little more.