Use util_log to initialize logger in tsdb-* tools (#12721)
#### What this PR does

When running these tools we would see a log like:

```
probe="internal message: if you see this, probably log initialization has gone wrong"
```

This comes from
[here](https://github.com/grafana/mimir/blob/main/pkg/util/log/slogadapter.go#L16),
and is mentioned in https://github.com/grafana/mimir/pull/10646 as a
side-effect of using a non `util_log` logger with `SlogFromGoKit`.

This PR changes the logger initialization to use `util_log` to avoid the
spurious line.

I don't think a changelog is required for this.

#### Which issue(s) this PR fixes or relates to

Fixes #

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.