storegateway: safeguard chunk parser (#12729)
#### What this PR does

As @dimitarvdimitrov brought in
https://github.com/grafana/mimir/pull/12693#discussion_r2352625595 the
`tsdb.EstimatedMaxChunkSize` is only an estimation, and we shouldn't use
it as the parser's cap. The PR updates the code to cap it with a more
relaxed value.

#### Which issue(s) this PR fixes or relates to

Follows https://github.com/grafana/mimir/pull/12693

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>
Co-authored-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>