Use conftest v0 flag (#12712)
#### What this PR does

Updates the build image to use the latest `conftest` binary. 

Updates the existing use of `conftest` to pass in the backwards
compatibility flag of `--rego-version=v0`.


[0.57.0](https://github.com/open-policy-agent/conftest/releases/tag/v0.57.0)
of conftest made a breaking change such that it assumes the default test
syntax is v1. Mimir tests use the older v0 format.

This PR allows for Mimir conftests to be run using a more recent
conftest version.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

N/A

---------

Co-authored-by: tcp13equals2 <tcp13equals2@users.noreply.github.com>