[release-2.17] Fix Timeseries CT reset before putting back to the pool (#12735)
#### What this PR does

Backporting to 2.17 a 1-line fix that was part of a bigger PR (#12652).
This fix an data corruption issue caused by memory pooling when a newer
version is rolled out and the experimental
-distributor.otel-created-timestamp-zero-ingestion-enabled feature is
enabled.

#### Which issue(s) this PR fixes or relates to

Related to #12652 

#### Checklist

- N/A Tests updated.
- N/A Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- N/A
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>