fix(distributor): NHCB bucket count validation wrong (#12741)
#### What this PR does

If the bucket limit is set for native histograms then no NHCB passes
validation due to a coding error.

Enabled bucket limit in test config to reproduce and fixed bug.

#### Which issue(s) this PR fixes or relates to

Related to RW2 compliance work.
https://github.com/prometheus/prometheus/issues/16944

#### Checklist

- [x] Tests updated.
- N/A Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- N/A
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>