Include more error information in query traces (#12734)
#### What this PR does

This PR improves the information logged in query traces when an error
occurs:
* MQE will now log whether the query succeeded or failed, and if it
failed, what the error was.
* Query-frontends will log the error returned by queriers, if any.

I've opted not to add a changelog entry given this only affects tracing.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.