distributor: Add truncate and drop options for long label values (#12627)
#### What this PR does

At the moment, a label value over the length limit always causes an
error. But most of the time, label values that long can just be
truncated or dropped altogether.

This PR adds an option to decide a strategy to handle a label value too
long error: truncate the label value, or drop the label altogether.

As a result, several timeseries that were before different may now be
the same. To avoid duplicated samples, the hash of the (original) label
value is set (at the end for "truncate", or fully replacing for "drop")
as the label value.

#### Which issue(s) this PR fixes or relates to

See backreferences.

#### Checklist

- [x] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [x]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Arve Knudsen <arve.knudsen@gmail.com>